requirements.txt
# 📄 requirements.txt
flask
flask-cors
python-dotenv
requests